Start using electron-vite
