+++ electron.vite.config.js
@@ -1,0 +1,8 @@
+export default {
+    main: {
+    },
+    preload: {
+    },
+    renderer: {
+    }
+}
