+++ package.json
@@ -2,10 +2,10 @@
   "name": "cms-launcher",
   "version": "0.1.9",
   "description": "A quick-start launcher for Drupal CMS.",
-  "main": "main.js",
+  "main": "./out/main/main.js",
   "scripts": {
-    "start": "electron .",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "electron-vite build",
+    "start": "electron-vite preview",
     "make": "electron-builder --publish never"
   },
   "author": "Drupal Association",
@@ -13,7 +13,9 @@
   "devDependencies": {
     "@playwright/test": "^1.52.0",
     "electron": "^36",
-    "electron-builder": "^26"
+    "electron-builder": "^26",
+    "electron-vite": "^3.1.0",
+    "vite": "^6.3.5"
   },
   "dependencies": {
     "get-port": "^7.1.0"
