+++ src/renderer/index.html
@@ -3,14 +3,8 @@
 <head>
   <meta charset="UTF-8" />
   <!-- https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP -->
-  <meta
-          http-equiv="Content-Security-Policy"
-          content="default-src 'self'; script-src 'self'"
-  />
-  <meta
-          http-equiv="X-Content-Security-Policy"
-          content="default-src 'self'; script-src 'self'"
-  />
+  <meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self'" />
+  <meta http-equiv="X-Content-Security-Policy" content="default-src 'self'; script-src 'self'" />
   <title>Drupal CMS</title>
   <link rel="stylesheet" type="text/css" href="./styles.css" />
 </head>
@@ -33,5 +27,5 @@
   </main>
 </div>
 </body>
-<script src="./renderer.js"></script>
+<script type="module" src="./renderer.js"></script>
 </html>
